migration_pjt/kdms_origin/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
apscheduler==3.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.121.1
fonttools==4.60.1
h11==0.16.0
httptools==0.7.1
idna==3.11
importlib-resources==6.5.2
iniconfig==2.1.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydantic==2.12.4
finance-datareader>=0.9.80
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tomli==2.3.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.9.8
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
zipp==3.23.0
